+++ app/src/hooks/useModal.ts
@@ -1,7 +1,7 @@
 import { useNavigation } from '@react-navigation/native';
 import { useCallback, useState } from 'react';
 
-import { ModalParams } from '../screens/Settings/ModalScreen';
+import { ModalParams } from '../screens/static/ModalScreen';
 
 export const useModal = (params: ModalParams) => {
   const [visible, setVisible] = useState(false);
