+++ app/src/screens/prove/QRCodeTroubleScreen.tsx
