+++ app/src/screens/prove/ViewFinderScreen.tsx
