+++ app/src/screens/settings/ShowRecoveryPhraseScreen.tsx
