+++ app/src/screens/static/LaunchScreen.tsx
