+++ app/src/screens/static/SplashScreen.tsx
